backend/requirements.txt
Django>=3.2,<4.2
djangorestframework>=3.12,<3.14
psycopg2-binary>=2.8
django-cors-headers>=3.5
django-environ==0.12.0
djangorestframework-simplejwt==5.3.1
Pillow>=10.4.0
celery==5.4.0
# Prayer times library
praytimes
celery
# Optional caching support
django-redis>=4.12 
# API documentation
drf-yasg>=1.20